[CH-3]:refactor

+++ models/chatmember.go
@@ -2,8 +2,8 @@ package models
 
 type ChatMember struct {
 	Base
-	UserID int `json:"user_id"`
-	RoomID int `json:"room_id"`
+	UserID int   `json:"user_id"`
+	RoomID int64 `json:"room_id"`
 }
 
 func (c ChatMember) TableName() string {
